+++ src/utilities/stringManipulations.ts
@@ -1,4 +1,5 @@
-import { FILE_HEADER_START_TOKEN, FILE_HEADER_END_TOKEN } from './consts';
+const FILE_HEADER_START_TOKEN = '#_';
+const FILE_HEADER_END_TOKEN = '_#';
 
 const makeHeaderLine = (headerStr: string) =>
   `${FILE_HEADER_START_TOKEN}${headerStr}${FILE_HEADER_END_TOKEN}\n`;
@@ -12,12 +13,6 @@ const extractHeaderLine = (headerLine: string) => {
   throw new Error('No descriptive headers found in .env');
 };
 
-const templateLabel = (env: string) => {
-  const upperEnv = env.toUpperCase();
-  const isProd = upperEnv.includes('PROD');
-  return `${isProd ? `$(issue-opened) ${upperEnv}` : env}`;
-};
-
 const capitalize = (str: string) => `${str.charAt(0).toUpperCase()}${str.slice(1)}`;
 
-export { makeHeaderLine, extractHeaderLine, templateLabel, capitalize };
+export { makeHeaderLine, extractHeaderLine, capitalize };